Add instructions and scaffolding for exercise A4.2
