+++ exercises/A4/2-local-storage-vec/Cargo.toml
